--- Pandator/Assets/InterruptGun/InterrutItem/Scripts/TestPlayerColorManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9b3b720a9aa5148fd82ba67681f7c79f
