+++ Pandator/Assets/NetGun/Net/Scripts/Net.cs
@@ -14,13 +14,13 @@ public class Net : MonoBehaviour
     private void Start()
     {
         animator = this.GetComponent<Animator>();
-        
+
         animator.SetTrigger("Capture");
     }
     private void Update()
     {
         animator.SetTrigger("Idle");
-        
+
         // 3秒後に消える
         time += Time.deltaTime;
         if(time > 3.0f)
@@ -35,26 +35,26 @@ public class Net : MonoBehaviour
         //         Destroy(gameObject);
         //     }
         // }
-        
+
     }
 
     private void OnTriggerEnter(Collider collision)
     {
         if(isCollision) return;
         GameObject Player = collision.gameObject;
-        
+
         if(Player.tag == "Player")
         {
             isCollision = true;
             //Netの中に入れる
             Player.GetComponent<SphereCollider>().isTrigger = true;
             //速度を0に
-            GetComponent<Rigidbody>().linearVelocity = Vector3.zero; 
+            GetComponent<Rigidbody>().linearVelocity = Vector3.zero;
             transform.position = Player.transform.position;
             transform.position += new Vector3(0, 1, 0);
             Player.GetComponent<StateManager>()?.SetAlive(false);
             // ここは視覚的にわかりやすいように色を変える処理を追加しているだけ
-            Player.GetComponent<TestPlayerColorManager>()?.ChangeColorBlack();
+            Player.GetComponent<PlayerColorManager>()?.ChangeColorBlack();
             Debug.Log("hit");
         }
     }
