+++ Pandator/Assets/Resources/Player/BirdPlayer.prefab
@@ -547,13 +547,13 @@ PrefabInstance:
     m_AddedComponents:
     - targetCorrespondingSourceObject: {fileID: 4549435062556395408, guid: ae6c99a9b9fae4ff887cf09947264e97, type: 3}
       insertIndex: -1
-      addedObject: {fileID: 1275615744167943154}
-    - targetCorrespondingSourceObject: {fileID: 4549435062556395408, guid: ae6c99a9b9fae4ff887cf09947264e97, type: 3}
-      insertIndex: -1
       addedObject: {fileID: 1955606003174083035}
     - targetCorrespondingSourceObject: {fileID: 4549435062556395408, guid: ae6c99a9b9fae4ff887cf09947264e97, type: 3}
       insertIndex: -1
       addedObject: {fileID: 5253004324674076423}
+    - targetCorrespondingSourceObject: {fileID: 4549435062556395408, guid: ae6c99a9b9fae4ff887cf09947264e97, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 5164755693211121568}
   m_SourcePrefab: {fileID: 100100000, guid: ae6c99a9b9fae4ff887cf09947264e97, type: 3}
 --- !u!4 &3621555379604008764 stripped
 Transform:
@@ -565,18 +565,6 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 4549435062556395408, guid: ae6c99a9b9fae4ff887cf09947264e97, type: 3}
   m_PrefabInstance: {fileID: 444553267706491414}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1275615744167943154
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4109967603769013638}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9b3b720a9aa5148fd82ba67681f7c79f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &1955606003174083035
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -591,7 +579,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   interruptedTime: 3
   interruptedSpeed: 2
-  playerColorManager: {fileID: 1275615744167943154}
+  playerColorManager: {fileID: 5164755693211121568}
   photonKeyMove: {fileID: 5823670346237709383}
   scoreManager: {fileID: 8028601576095073563}
   parentObject: {fileID: 0}
@@ -616,6 +604,18 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 0.43800396
   m_Center: {x: -0.053215742, y: 0.4887637, z: 0.07171863}
+--- !u!114 &5164755693211121568
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4109967603769013638}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 45b7416fc37c44b30b73190a94051c6b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &5338401572920123032 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 5493988184530889870, guid: ae6c99a9b9fae4ff887cf09947264e97, type: 3}
