+++ Pandator/Assets/Scenes/VR/VRScene.prefab
@@ -172,7 +172,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &1295255931325412875
 Transform:
   m_ObjectHideFlags: 0
